requirements.txt
amqp==5.0.6
asgiref==3.3.1
billiard==3.6.4.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cycler==0.10.0
decorator==4.4.2
Django==3.1.7
django-background-tasks==1.2.5
django-compat==1.0.15
importlib-metadata==3.10.0
kiwisolver==1.3.1
kombu==5.0.2
matplotlib==3.3.4
networkx==2.5
numpy==1.20.1
Pillow==8.1.2
prompt-toolkit==3.0.18
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
